1. 模型文件夹改为以model开头 2. 增加20260317模型 3. 支持3张jpg的训练模式

+++ utils/generateOnePicDataset_npy2jpg.py
@@ -10,34 +10,35 @@ def build_onepic_dataset(dataset_root, vid_name, frame_id, repeat_num, output_ro
     # -------------------------
     output_root = os.path.join(dataset_root, output_root)
     if os.path.exists(output_root):
-        print(f"错误: {output_root} 已存在，请手动删除后再运行")
-        sys.exit(1)
+        print(f"警告: {output_root} 已存在，请手动删除后再运行")
 
     # 创建目录
     mot_dst = os.path.join(output_root, "mot")
-    npy_dst = os.path.join(output_root, "npy", vid_name)
+    npy2jpg_dst = os.path.join(output_root, "npy2jpg", vid_name)
     train_dst = os.path.join(output_root, "train")
     test_dst = os.path.join(output_root, "test")
 
-    os.makedirs(mot_dst)
-    os.makedirs(npy_dst)
-    os.makedirs(train_dst)
-    os.makedirs(test_dst)
+    os.makedirs(mot_dst, exist_ok=True)
+    os.makedirs(npy2jpg_dst, exist_ok=True)
+    os.makedirs(train_dst, exist_ok=True)
+    os.makedirs(test_dst, exist_ok=True)
 
     # -------------------------
     # 源路径
     # -------------------------
     mot_src = os.path.join(dataset_root, "mot")
-    npy_src = os.path.join(dataset_root, "npy")
+    npy2jpg_src = os.path.join(dataset_root, "npy2jpg")
 
     txt_src = os.path.join(mot_src, f"{vid_name}.txt")
-    npy_frame_src = os.path.join(npy_src, vid_name, f"{frame_id:06d}.npy")
+    npy2jpg_frame_src = [os.path.join(npy2jpg_src, vid_name, f"{frame_id:06d}_p1.jpg"),
+                        os.path.join(npy2jpg_src, vid_name, f"{frame_id:06d}_p2.jpg"),
+                        os.path.join(npy2jpg_src, vid_name, f"{frame_id:06d}_p3.jpg")]
 
     if not os.path.exists(txt_src):
         raise FileNotFoundError(txt_src)
 
-    if not os.path.exists(npy_frame_src):
-        raise FileNotFoundError(npy_frame_src)
+    if not all(os.path.exists(frame_src) for frame_src in npy2jpg_frame_src):
+        raise FileNotFoundError(npy2jpg_frame_src)
 
     # -------------------------
     # 1 读取txt筛选帧
@@ -74,18 +75,23 @@ def build_onepic_dataset(dataset_root, vid_name, frame_id, repeat_num, output_ro
     # 3 复制npy
     # -------------------------
     for i in range(repeat_num):
-        dst_name = f"{i+1:06d}.npy"
-        dst_path = os.path.join(npy_dst, dst_name)
-        shutil.copy(npy_frame_src, dst_path)
+        for part_idx in range(3):
+            src_name = npy2jpg_frame_src[part_idx]
+            dst_name = f"{i+1:06d}_p{part_idx+1}.jpg"
+            dst_path = os.path.join(npy2jpg_dst, dst_name)
+            shutil.copy(src_name, dst_path)
 
     # -------------------------
     # 4 创建软链接
     # -------------------------
-    os.symlink("../mot", os.path.join(train_dst, "mot"))
-    os.symlink("../npy", os.path.join(train_dst, "npy"))
-
-    os.symlink("../mot", os.path.join(test_dst, "mot"))
-    os.symlink("../npy", os.path.join(test_dst, "npy"))
+    if not os.path.exists(os.path.join(train_dst, "mot")):
+        os.symlink("../mot", os.path.join(train_dst, "mot"))
+    if not os.path.exists(os.path.join(train_dst, "npy2jpg")):
+        os.symlink("../npy2jpg", os.path.join(train_dst, "npy2jpg"))
+    if not os.path.exists(os.path.join(test_dst, "mot")):
+        os.symlink("../mot", os.path.join(test_dst, "mot"))
+    if not os.path.exists(os.path.join(test_dst, "npy2jpg")):
+        os.symlink("../npy2jpg", os.path.join(test_dst, "npy2jpg"))
 
     print("OnePic 数据集生成完成")
     print("输出目录:", output_root)
